+++ port/pic32-harmony/create_examples.py
@@ -52,7 +52,7 @@ print("Creating examples in apps/btstack")
 for file in os.listdir(examples_embedded):
     if not file.endswith(".c"):
         continue
-    if file.endswith("sco_demo_util.c"):
+    if file in ['panu_demo.c', 'sco_demo_util.c', 'ant_test.c', 'pan_lwip_http_server.c']:
         continue
     example = file[:-2]