+++ port/wiced-h5/create_examples.py
@@ -110,6 +110,8 @@ print("\nCreating examples in apps/btstack:")
 for file in os.listdir(examples_embedded):
     if not file.endswith(".c"):
         continue
+    if file in ['panu_demo.c', 'sco_demo_util.c', 'ant_test.c', 'pan_lwip_http_server.c']:
+        continue
     example = file[:-2]
 
     # create folder